refactor: Move route-specific stores to route folders
Colocate stores with their routes instead of creating separate
feature folders since these stores are only used by their
respective routes.
