--- src/client/features/manage-plan/index.ts
@@ -1,15 +1,0 @@
-// Export store and derived hooks
-export {
-    useManagePlanStore,
-    useManagePlanActiveTab,
-    useManagePlanFilterMuscle,
-    useManagePlanFilterType,
-    useManagePlanFilterSource,
-    useSetManagePlanActiveTab,
-    useSetManagePlanFilterMuscle,
-    useSetManagePlanFilterType,
-    useSetManagePlanFilterSource,
-} from './store';
-
-// Export types
-export type { ManagePlanState, ManagePlanTab, FilterSource } from './types';
