--- src/client/features/progress/index.ts
@@ -1,11 +1,0 @@
-// Export store and derived hooks
-export {
-    useProgressStore,
-    useProgressActiveTab,
-    useProgressDateRange,
-    useSetProgressActiveTab,
-    useSetProgressDateRange,
-} from './store';
-
-// Export types
-export type { ProgressState, ProgressTab, DateRange } from './types';
