--- src/client/features/reports/index.ts
@@ -1,15 +1,0 @@
-// Export store and derived hooks
-export {
-    useReportsStore,
-    useReportsTypeFilter,
-    useReportsStatusFilter,
-    useReportsSortOrder,
-    useReportsViewMode,
-    useSetReportsTypeFilter,
-    useSetReportsStatusFilter,
-    useSetReportsSortOrder,
-    useSetReportsViewMode,
-} from './store';
-
-// Export types
-export type { ReportsState, ReportsViewMode, ReportsSortOrder } from './types';
