+++ src/client/routes/Reports/Reports.tsx
@@ -6,7 +6,7 @@
 
 import { useState, useEffect, useRef } from 'react';
 import { useReports, useUpdateReportStatus, useDeleteReport, useDeleteAllReports } from './hooks';
-import { useReportsStore } from '@/client/features/reports';
+import { useReportsStore } from './store';
 import { Card, CardContent, CardHeader, CardTitle } from '@/client/components/ui/card';
 import { Button } from '@/client/components/ui/button';
 import { Badge } from '@/client/components/ui/badge';
